Redo architecture and accept user inputs
