+++ src/validate_keyword.go
@@ -1,0 +1,13 @@
+package source
+
+import (
+	"strings"
+)
+
+func ValidateKeyword(rawKeyword string) (string, error) {
+	// Trim newline characters from the input
+	rawKeyword = strings.TrimSpace(rawKeyword)
+
+	validKeyword := strings.ReplaceAll(rawKeyword, " ", "+")
+	return validKeyword, nil
+}
